requirements.txt
pyyaml==3.12
sewer==0.5.2
prompt_toolkit==1.0.15
GitPython==2.1.9
python-nginx==1.4.1
blessings==1.6.1
